+++ src/errors/runner.mjs
@@ -1,6 +1,8 @@
+import { EOL } from 'os'
+
 export class RunnerError extends Error {
   // eslint-disable-next-line space-before-function-paren, no-useless-constructor
   constructor(message) {
-    super(message)
+    super('Runner error:' + EOL + message)
   }
 }
