+++ src/errors/timeout.mjs
@@ -1,9 +1,17 @@
 /* eslint-disable promise/param-names */
 /* eslint-disable space-before-function-paren */
 /* eslint-disable prefer-promise-reject-errors */
-export class TestTimeoutError extends Error {
+import { EOL } from 'os'
+
+export class TimeoutError extends Error {
   constructor(timeout) {
-    super(`Test timed out after ${timeout}ms`)
+    super(
+      'Runner error:' +
+        EOL +
+        `Exceeded timeout of ${timeout} ms for a test.` +
+        EOL +
+        'Use { timeout: 10_000} to increase the timeout value, if this is a long-running test.'
+    )
     this.timeout = timeout
   }
