+++ src/reporters/reporter.mjs
@@ -1,6 +1,6 @@
 import { getConfig } from '../config.mjs'
 import { REPORTER_TYPE } from '../constants.mjs'
-import { RunnerError } from '../runnerError.mjs'
+import { RunnerError } from '../errors/runner.mjs'
 import { prepareHTMLReport } from './htmlReporter.mjs'
 import { prepareJsonReport } from './jsonReporter.mjs'
